removed deprecated read_stream FB that FB refuses to grant anymore. The idea was to use it at some point to allow you to read your FB feed via Temboz, and thus benefit from it's filtering capabilities, but it ended up never being implemented. Needless to say, FB's action is user-hostile, they don't want third-party apps to bypass the official FB UIs and the advertising that goes with them.
More details from:
https://developers.facebook.com/docs/facebook-login/permissions#reference-read_stream

read_stream

This permission is only available for apps using Graph API version v2.3 or older. See user_posts as a possible alternative.

Provides access to read the posts in a person's News Feed, or the posts on their Profile.

Review
If your app requests this permission Facebook will have to review how your app uses it.

Limited Use

This permission is granted to apps building a Facebook-branded client on platforms where Facebook is not already available. For example, Android and iOS apps will not be approved for this permission. In addition, Web, Desktop, in-car and TV apps will not be granted this permission.